requirements.txt
streamlit>=1.36.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pypdf>=4.2.0
python-docx>=1.1.0
pandas>=2.2.2
numpy>=1.26.4
pytest>=8.2.0


